Another small refactor
